requirements.txt
vcon
fastapi
uvicorn
pydantic
httpx
pypdf
Pillow
sqlalchemy
python-dotenv
psycopg2-binary